requirements.txt
streamlit==1.32.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
face-recognition==1.3.0
face-recognition-models==0.3.0
pandas==2.2.2
streamlit-webrtc==0.47.4
plotly==5.22.0
Pillow==10.3.0
torch==2.3.0
torchvision==0.18.0
timm==1.0.0
dlib==19.24.4
cmake==3.29.4

